[mac] Breakout DynamicFill into reusable
